requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
boto3==1.28.57
botocore==1.31.57
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
executing==2.0.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.4
ipython==8.16.0
iso8601==1.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
numpy==1.26.0
openpyxl==3.1.2
orjson==3.9.15
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg-binary==3.1.12
pure-eval==0.2.2
pydantic==2.6.3
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.16.1
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.36.3
tortoise-orm==0.20.0
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.9.0
urllib3==1.26.16
uvicorn==0.27.1
watchfiles==0.21.0
wcwidth==0.2.8
websockets==12.0
